Ordeno las paginas en subcarpetas según las sección a la que corresponde
