+++ src/app/app.module.ts
@@ -13,27 +13,27 @@ import { EscanerDniPage } from '../pages/registro/escaner-dni/escaner-dni';
 import { RegistroPersonalDataPage } from '../pages/registro/personal-data/personal-data';
 import { RegistroUserDataPage } from '../pages/registro/user-data/user-data';
 import { LoginPage } from '../pages/login/login';
-import { NavbarPage } from '../pages/navbar/navbar';
+import { NavbarPage } from '../components/navbar/navbar';
 import { VerificaCodigoPage } from '../pages/registro/verifica-codigo/verifica-codigo';
 import { BienvenidaPage } from '../pages/bienvenida/bienvenida';
 import { WaitingValidationPage } from '../pages/registro/waiting-validation/waiting-validation';
 import { ProfilePacientePage } from '../pages/profile/paciente/profile-paciente';
 import { ProfileAccountPage } from '../pages/profile/account/profile-account';
 import { EditorPacientePage } from '../pages/profile/editor-paciente/editor-paciente';
-import { OrganizacionesPage } from '../pages/organizaciones/organizaciones';
+import { OrganizacionesPage } from '../pages/login/organizaciones/organizaciones';
 import { AgendasPage } from '../pages/profesional/agendas/agendas';
 import { NumerosUtilesPage } from '../pages/datos-utiles/numeros-emergencia/numeros-utiles';
 import { FarmaciasTurnoPage } from '../pages/datos-utiles/farmacias-turno/farmacias-turno';
 import { FeedNoticiasPage } from '../pages/datos-utiles/feed-noticias/feed-noticias';
 import { VacunasPage } from '../pages/vacunas/vacunas';
-import { MapPage } from '../pages/centros-salud/map/map';
-import { ListPage } from '../pages/centros-salud/list/list';
-import { CentrosSaludPage } from '../pages/centros-salud/centros-salud';
+import { MapPage } from '../pages/datos-utiles/centros-salud/map/map';
+import { ListPage } from '../pages/datos-utiles/centros-salud/list/list';
+import { CentrosSaludPage } from '../pages/datos-utiles/centros-salud/centros-salud';
 import { DondeVivoDondeTrabajoPage } from '../pages/profile/paciente/donde-vivo-donde-trabajo/donde-vivo-donde-trabajo';
-import { FaqPage } from '../pages/faq/faq';
+import { FaqPage } from '../pages/datos-utiles/faq/faq';
 import { HistoriaDeSaludPage } from '../pages/historia-salud/historia-salud';
 import { InformacionValidacionPage } from '../pages/registro/informacion-validacion/informacion-validacion';
-import { RecuperarPasswordPage } from '../pages/profile/recuperar-password/recuperar-password';
+import { RecuperarPasswordPage } from '../pages/registro/recuperar-password/recuperar-password';
 
 // Plugins
 import { StatusBar } from '@ionic-native/status-bar';
