+++ src/config.ts
@@ -2,11 +2,10 @@ export default {
   REMEMBER_SESSION: true,
   APP_VERSION: 4,
   MAP_KEY: 'AIzaSyAqYUabuhvLRq3c7CYDH6aTu5NHIpTTFQ4',
-  // API_URL: 'https://andesapps.hospitalneuquen.org.ar/api/'
-
-  API_URL: 'http://192.168.0.116:3002/api/',
-  // API_URL: 'http://mongolito.hospitalneuquen.org.ar/api/',
+  API_URL: 'https://andesapps.hospitalneuquen.org.ar/api/'
   // API_URL: 'http://d-testing.hospitalneuquen.org.ar/api/'
+  // API_URL: 'http://192.168.0.116:3002/api/',
+  // API_URL: 'http://mongolito.hospitalneuquen.org.ar/api/',
   // API_URL: 'http://192.168.0.13:3002/api/'
   // API_URL: 'http://192.168.4.8:3002/api/'
 }
