--- src/interfaces/usuario.interface.ts
@@ -1,13 +1,0 @@
-export interface Usuario {
-  nombre: string;
-  appelido: string;
-  email: string;
-  dni: string;
-  nacionalidad: string;
-  sexo: string;
-  genero: string;
-  telefono: string;
-  password: string;
-  confirmPassword: string;
-  codigoVerificacion: string;
-}
