+++ src/pages/datos-utiles/centros-salud/map/map.ts
@@ -1,7 +1,7 @@
 import { NavController, NavParams, Platform, AlertController } from 'ionic-angular';
 import { Component, ElementRef, ViewChild, NgZone } from '@angular/core';
-import { LocationsProvider } from '../../../providers/locations/locations';
-import { GoogleMapsProvider } from '../../../providers/google-maps/google-maps';
+import { LocationsProvider } from '../../../../providers/locations/locations';
+import { GoogleMapsProvider } from '../../../../providers/google-maps/google-maps';
 
 import { Geolocation } from '@ionic-native/geolocation';
 import { NativeGeocoder, NativeGeocoderReverseResult } from '@ionic-native/native-geocoder';
@@ -199,7 +199,9 @@ export class MapPage {
   }
 
   ngOnDestroy() {
-    this.geoSubcribe.unsubscribe();
+    if (this.geoSubcribe) {
+      this.geoSubcribe.unsubscribe();
+    }
   }
 
   /* Se comenta hasta que se defina bien como va a funcionar*/
