+++ src/pages/profile/account/profile-account.ts
@@ -4,7 +4,6 @@ import { FormGroup } from '@angular/forms';
 import { NavController, NavParams, LoadingController, MenuController } from 'ionic-angular';
 import { AlertController } from 'ionic-angular';
 import { AuthProvider } from '../../../providers/auth/auth';
-import { Usuario } from '../../../interfaces/usuario.interface';
 import { Storage } from '@ionic/storage';
 import { ToastProvider } from '../../../providers/toast';
 
@@ -13,7 +12,6 @@ import { ToastProvider } from '../../../providers/toast';
   templateUrl: 'profile-account.html',
 })
 export class ProfileAccountPage {
-  public usuario: Usuario;
   loading: any;
   mostrarMenu: boolean = true;
   fase: number = 1;
