+++ src/pages/profile/editor-paciente/editor-paciente.ts
@@ -3,7 +3,6 @@ import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { NavController, NavParams, LoadingController, MenuController } from 'ionic-angular';
 import { AlertController } from 'ionic-angular';
 import { AuthProvider } from '../../../providers/auth/auth';
-import { Usuario } from '../../../interfaces/usuario.interface';
 import { Storage } from '@ionic/storage';
 import { PacienteProvider } from '../../../providers/paciente';
 import { ToastProvider } from '../../../providers/toast';
@@ -13,7 +12,6 @@ import { ToastProvider } from '../../../providers/toast';
   templateUrl: 'editor-paciente.html',
 })
 export class EditorPacientePage {
-  public usuario: Usuario;
   public paciente: any;
 
   loading: any;
