+++ src/pages/registro/recuperar-password/recuperar-password.scss
